% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELMEROANCOVA.R
\name{ana_covarianza}
\alias{ana_covarianza}
\title{Analisis de covarianza.}
\usage{
ana_covarianza(var_resp, var_explic)
}
\arguments{
\item{var_resp}{(vector) datos respuesta.}

\item{var_explic}{(vector) datos explicativos.}

\item{dtf}{el data frame que contiene a los datos.}

\item{bloque1:ny}{(vector) datos repuesta por bloque.}

\item{bloque1:nx}{(vector) datos explicativos por bloque.}

\item{totvar_explic1:nx}{suma de los datos explcativos por bloque.}

\item{totvar_resp1:nx}{suma de los datos respuesta por bloque.}

\item{trat1:ny}{(vector) datos repuesta por tratamiento.}

\item{trat1:nx}{(vector) datos explicativos por tratamiento.}

\item{totvar_explictrat1:nx}{suma de los datos explcativos por tratamiento.}

\item{totvar_resptrat1:ny}{suma de los datos respuesta por tratamiento.}

\item{ajust_model}{modelo ajustado.}

\item{coeficientes}{coeficiente del modelo ajustado.}
}
\value{
el resultado del analisis de covarianza y el modelo ajustado de regresion lineal simple.
}
\description{
Realiza analisis de covarianza con una sola variable explicativa.
}
